SQUASH ac97: be more greedy at the start of decoding
When the decoder finds the very first edge to be falling, it has missed
the start of that bit time, but can grab a valid SYNC level, and thus
could find an earlier start of a frame at the beginning of the input
stream, which a previous implementation would have missed.

Rename a few variables for readability while we are here. Use better
Python idioms for the edge detection "shift register".