HACK ac97: add binary annotations (frame bits, 16bit audio samples)
Add support for binary annotations in the AC'97 decoder. Provide four
binary classes: output and input each for the frames' bits and for the
data in those slots which pass the dummy handler -- the latter shall
result in the extraction of all audio channels.

This is a rough hack. Bit to byte field conversion might need some more
attention. Frame bits are only sent for the previous frame, when the
start of another frame is seen.